Encounter events.
