+++ NWN.Anvil/src/main/API/Events/Game/EncounterEvents/EncounterEvents.OnExit.cs
@@ -1,0 +1,39 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwEncounter"/> triggers.
+  /// </summary>
+  public static partial class EncounterEvents
+  {
+    [GameEvent(EventScriptType.EncounterOnObjectExit)]
+    public sealed class OnExit : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwEncounter"/> that was exited.
+      /// </summary>
+      public NwEncounter Encounter { get; } = NWScript.OBJECT_SELF.ToNwObject<NwEncounter>();
+
+      NwObject IEvent.Context
+      {
+        get => Encounter;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwEncounter
+  {
+    /// <inheritdoc cref="EncounterEvents.OnExit"/>
+    public event Action<EncounterEvents.OnExit> OnExit
+    {
+      add => EventService.Subscribe<EncounterEvents.OnExit, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<EncounterEvents.OnExit, GameEventFactory>(this, value);
+    }
+  }
+}
