+++ NWN.Anvil/src/main/API/Events/Game/EncounterEvents/EncounterEvents.OnUserDefined.cs
@@ -1,0 +1,45 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwEncounter"/> triggers.
+  /// </summary>
+  public static partial class EncounterEvents
+  {
+    [GameEvent(EventScriptType.EncounterOnUserDefinedEvent)]
+    public sealed class OnUserDefined : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwEncounter"/> associated with this user defined event.
+      /// </summary>
+      public NwEncounter Encounter { get; } = NWScript.OBJECT_SELF.ToNwObject<NwEncounter>();
+
+      NwObject IEvent.Context
+      {
+        get => Encounter;
+      }
+
+      public static void Signal(NwEncounter encounter, int eventId)
+      {
+        Event nwEvent = NWScript.EventUserDefined(eventId);
+        NWScript.SignalEvent(encounter, nwEvent);
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwEncounter
+  {
+    /// <inheritdoc cref="EncounterEvents.OnUserDefined"/>
+    public event Action<EncounterEvents.OnUserDefined> OnUserDefined
+    {
+      add => EventService.Subscribe<EncounterEvents.OnUserDefined, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<EncounterEvents.OnUserDefined, GameEventFactory>(this, value);
+    }
+  }
+}
